mock other components; c/displayPanel
There will be times when the LWC you are working with imports another LWC that is not in the local LWC directory of your Salesforce DX project. Without a stub they will cause Jest to fail.

displayPanel.js: Notice that there is an api decorator for each parameter passed in the component call.

https://trailhead.salesforce.com/content/learn/modules/test-lightning-web-components/mock-other-components?trail_id=build-lightning-web-components